+++ src/main/kotlin/gather/FileGenerationListener.kt
@@ -1,0 +1,12 @@
+package gather
+
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.util.messages.Topic
+
+interface FileGenerationListener {
+    companion object {
+        val TOPIC = Topic.create("Solution File Generation", FileGenerationListener::class.java)
+    }
+
+    fun onGenerated(file: VirtualFile)
+}